fix: tighten chart axis spacing
