+++ frontend/src/accounts/detail/components/BalanceChartCard.tsx
@@ -1,6 +1,5 @@
-import { useEffect, useMemo, useState } from 'react'
+import { useMemo, useState } from 'react'
 import { TrendingDown, TrendingUp } from 'lucide-react'
-import { motion, useReducedMotion } from 'motion/react'
 import {
   Area,
   AreaChart,
@@ -14,10 +13,19 @@ import { useAccountSnapshots, type Account } from '@/api/accounts'
 import { TimeRangeSelector, type TimeRangeSelectorOption } from '@/components/TimeRangeSelector'
 import { formatCurrency } from '@/utils/formatCurrency'
 import {
+  BalanceChartModeSelector,
+  type BalanceChartMode,
+} from '@/accounts/detail/components/BalanceChartModeSelector'
+import {
   RANGE_CONFIG,
-  TIME_SELECTOR_SPRING,
   type BalanceRange,
 } from '@/accounts/detail/constants/accountDetail'
+import {
+  calendarDateMs,
+  formatSignedBalanceCurrency,
+  formatUtcAxisDate,
+  getBalanceXAxisTicks,
+} from '@/accounts/detail/utils/balanceChartAxis'
 import {
   buildChartSeries,
   rezeroSeriesToPeriod,
@@ -31,21 +39,7 @@ const BALANCE_RANGE_OPTIONS: TimeRangeSelectorOption<BalanceRange>[] = [
   { value: '90D', label: '90D', description: 'Last 90 days' },
   { value: '1Y', label: '1Y', description: 'Last year' },
 ]
-type BalanceChartMode = 'balance' | 'change'
-const BALANCE_CHART_MODE_OPTIONS: Array<{ value: BalanceChartMode; label: string }> = [
-  { value: 'balance', label: 'Balance' },
-  { value: 'change', label: 'Change' },
-]
-const BALANCE_AXIS_TICK_COUNT_BY_RANGE: Record<BalanceRange, number> = {
-  '7D': 7,
-  '30D': 6,
-  '90D': 6,
-  '1Y': 6,
-}
-const DAY_MS = 24 * 60 * 60 * 1000
-const BALANCE_AXIS_EDGE_PADDING_DESKTOP_PX = 28
-const BALANCE_AXIS_EDGE_PADDING_MOBILE_PX = 4
-const BALANCE_CHART_COMPACT_QUERY = '(max-width: 1049px)'
+const BALANCE_AXIS_EDGE_PADDING_PX = 4
 
 type AxisTickProps = {
   x?: number | string
@@ -53,58 +47,6 @@ type AxisTickProps = {
   payload?: {
     value: number | string
   }
-}
-
-function joinClassNames(...classNames: Array<string | undefined | false>) {
-  return classNames.filter(Boolean).join(' ')
-}
-
-function useMediaQuery(query: string) {
-  const [matches, setMatches] = useState(() =>
-    typeof window !== 'undefined' ? window.matchMedia(query).matches : false,
-  )
-
-  useEffect(() => {
-    const mediaQuery = window.matchMedia(query)
-    const handleChange = () => setMatches(mediaQuery.matches)
-    handleChange()
-    mediaQuery.addEventListener('change', handleChange)
-    return () => mediaQuery.removeEventListener('change', handleChange)
-  }, [query])
-
-  return matches
-}
-
-function calendarDateMs(d: Date): number {
-  return Date.UTC(d.getFullYear(), d.getMonth(), d.getDate())
-}
-
-function getXAxisTicks(fromDate: Date, toDate: Date, range: BalanceRange): number[] {
-  const startMs = calendarDateMs(fromDate)
-  const endMs = calendarDateMs(toDate)
-  const tickCount = BALANCE_AXIS_TICK_COUNT_BY_RANGE[range]
-  if (tickCount <= 1 || startMs >= endMs) return [startMs]
-
-  const totalDays = Math.round((endMs - startMs) / DAY_MS)
-  return [...new Set(Array.from({ length: tickCount }, (_, index) => {
-    if (index === 0) return startMs
-    if (index === tickCount - 1) return endMs
-    const dayOffset = Math.round((totalDays * index) / (tickCount - 1))
-    return startMs + dayOffset * DAY_MS
-  }))]
-}
-
-function formatUtcAxisDate(value: number): string {
-  return new Intl.DateTimeFormat('en-US', {
-    month: 'short',
-    day: 'numeric',
-    timeZone: 'UTC',
-  }).format(new Date(value))
-}
-
-function formatSignedCurrency(amount: number, currency: string): string {
-  if (amount === 0) return formatCurrency(amount, currency)
-  return `${amount > 0 ? '+' : '−'}${formatCurrency(Math.abs(amount), currency)}`
 }
 
 function BalanceXAxisTick({
@@ -138,70 +80,9 @@ function BalanceXAxisTick({
   )
 }
 
-function BalanceChartModeSelector({
-  value,
-  onChange,
-  className,
-}: {
-  value: BalanceChartMode
-  onChange: (value: BalanceChartMode) => void
-  className?: string
-}) {
-  const shouldReduceMotion = useReducedMotion()
-  const activeIndex = Math.max(
-    BALANCE_CHART_MODE_OPTIONS.findIndex((option) => option.value === value),
-    0,
-  )
-
-  return (
-    <div
-      className={joinClassNames(
-        'app-segmented-control app-segmented-control-compact app-balance-chart-mode-selector min-w-0',
-        className,
-      )}
-      role="tablist"
-      aria-label="Balance chart view"
-    >
-      <motion.span
-        className="pointer-events-none absolute rounded-md"
-        aria-hidden
-        style={{
-          top: '0.125rem',
-          bottom: '0.125rem',
-          left: '0.125rem',
-          width: `calc((100% - 0.25rem) / ${BALANCE_CHART_MODE_OPTIONS.length})`,
-          background: 'var(--app-accent-soft)',
-          border: '1px solid var(--app-accent-border)',
-        }}
-        animate={{ x: `${activeIndex * 100}%` }}
-        transition={shouldReduceMotion ? { duration: 0 } : TIME_SELECTOR_SPRING}
-      />
-      {BALANCE_CHART_MODE_OPTIONS.map((option) => {
-        const active = option.value === value
-        return (
-          <button
-            key={option.value}
-            type="button"
-            role="tab"
-            aria-selected={active}
-            onClick={() => onChange(option.value)}
-            className={joinClassNames(
-              'app-segmented-option app-segmented-option-compact relative z-10 flex-1 px-3',
-              active && 'app-segmented-option-active',
-            )}
-          >
-            {option.label}
-          </button>
-        )
-      })}
-    </div>
-  )
-}
-
 export default function BalanceChartCard({ account }: { account: Account }) {
   const [range, setRange] = useState<BalanceRange>('30D')
   const [chartMode, setChartMode] = useState<BalanceChartMode>('balance')
-  const compactChart = useMediaQuery(BALANCE_CHART_COMPACT_QUERY)
 
   // Derive the snapshot query window from the selected range.
   const { fromDate, toDate, granularity } = useMemo(() => {
@@ -229,7 +110,10 @@ export default function BalanceChartCard({ account }: { account: Account }) {
   const chartDataKey = chartMode === 'balance' ? 'balance' : 'periodBalance'
   const axisStartMs = calendarDateMs(fromDate)
   const axisEndMs = calendarDateMs(toDate)
-  const xAxisTicks = useMemo(() => getXAxisTicks(fromDate, toDate, range), [fromDate, toDate, range])
+  const xAxisTicks = useMemo(
+    () => getBalanceXAxisTicks(fromDate, toDate, range),
+    [fromDate, toDate, range],
+  )
   const seriesByDateMs = useMemo(
     () => new Map(series.map((point) => [point.dateMs, point])),
     [series],
@@ -259,9 +143,6 @@ export default function BalanceChartCard({ account }: { account: Account }) {
       ? 'var(--app-positive)'
       : 'var(--app-negative)'
   const chartLineColor = chartMode === 'change' && periodDelta !== null ? deltaColor : lineColor
-  const axisEdgePadding = compactChart
-    ? BALANCE_AXIS_EDGE_PADDING_MOBILE_PX
-    : BALANCE_AXIS_EDGE_PADDING_DESKTOP_PX
 
   return (
     <section
@@ -269,7 +150,11 @@ export default function BalanceChartCard({ account }: { account: Account }) {
     >
       <div className="mb-4 flex flex-wrap items-center justify-between gap-3">
         <p className="app-label">Current Balance</p>
-        <BalanceChartModeSelector value={chartMode} onChange={setChartMode} className="w-[9.5rem] shrink-0 min-[750px]:hidden" />
+        <BalanceChartModeSelector
+          value={chartMode}
+          onChange={setChartMode}
+          className="w-[9.5rem] shrink-0 min-[750px]:hidden"
+        />
         <div className="ml-auto hidden items-center gap-2 min-[750px]:flex">
           <BalanceChartModeSelector value={chartMode} onChange={setChartMode} className="w-[9.5rem]" />
           <TimeRangeSelector
@@ -326,7 +211,15 @@ export default function BalanceChartCard({ account }: { account: Account }) {
           </div>
         ) : (
           <ResponsiveContainer width="100%" height="100%">
-            <AreaChart data={chartSeries} margin={{ top: 18, right: axisEdgePadding, bottom: 0, left: axisEdgePadding }}>
+            <AreaChart
+              data={chartSeries}
+              margin={{
+                top: 18,
+                right: BALANCE_AXIS_EDGE_PADDING_PX,
+                bottom: 0,
+                left: BALANCE_AXIS_EDGE_PADDING_PX,
+              }}
+            >
               <defs>
                 <linearGradient id={`balanceFill-${account.id}`} x1="0" y1="0" x2="0" y2="1">
                   <stop offset="0%" stopColor={chartLineColor} stopOpacity={0.22} />
@@ -362,7 +255,7 @@ export default function BalanceChartCard({ account }: { account: Account }) {
                 formatter={(value) => [
                   chartMode === 'balance'
                     ? formatCurrency(Number(value), account.currency)
-                    : formatSignedCurrency(Number(value), account.currency),
+                    : formatSignedBalanceCurrency(Number(value), account.currency),
                   chartMode === 'balance' ? 'Balance' : 'Change',
                 ]}
               />
